Split remote_build controller into Offloader + Receiver siblings
The 2837-line ``RemoteBuildController`` mixed two distinct roles
that share nothing but a ``DeviceBuilder`` reference: the
*outbound* side (pair flow, peer-link clients, submit_job /
cancel_job / download_artifacts, mDNS browse) and the *inbound*
side (peer-link session registry, pair inbox, queue_status
fan-out, identity rotate, cleanup sweep). A pre-split analysis
confirmed only 10 cross-role method references — all in
lifecycle (``start`` / ``stop``) and one in the mDNS browse
callback that routes naturally to the offloader.

Replace the single class with two siblings exposed on
``DeviceBuilder``:

* ``self.remote_build_offloader: OffloaderController`` — owns
  47 offloader methods plus mDNS browse (7), peer-link resolver
  (2), and the X25519 + dashboard_id loader. Lives in
  ``controllers/remote_build/offloader.py``.
* ``self.remote_build_receiver: ReceiverController`` — owns
  34 receiver methods plus identity ``get`` / ``rotate``. Lives
  in ``controllers/remote_build/receiver.py``.
* ``controllers/remote_build/_shared.py`` — thin module with
  the ``drain_tasks`` free helper both siblings need.

The umbrella ``RemoteBuildController`` class is gone — no
delegating facade. Each sibling has its own ``__init__`` /
``start`` / ``stop`` / ``_tasks`` / ``_listeners`` /
``_shutdown_callbacks``. ``peer_link.py`` and ``job_fanout.py``
take ``ReceiverController`` as their controller arg;
``firmware/controller.py`` and ``firmware/remote_runner.py``
reach into ``remote_build_offloader`` for
``build_scheduler_snapshot`` / ``get_pairing`` /
``_lookup_open_peer_link_client``.

Tests use a small ``RemoteBuildTestHandles`` shim in
``tests/conftest.py`` that bundles both siblings with
``__getattr__`` / ``__setattr__`` forwarding, so existing test
code addressing both sides through one controller keeps
working. New tests should reach ``handles.offloader`` /
``handles.receiver`` directly.

All 3117 tests pass; ruff + ruff-format + the rest of the
pre-commit chain are clean. No behaviour change — the split
moves method bodies verbatim and partitions ``__init__`` /
``start`` / ``stop`` along the role boundary the methods
already followed.
