+++ tests/test_remote_build_peer_link_client.py
@@ -36,9 +36,9 @@ from cryptography.hazmat.primitives.asymmetric.x25519 import X25519PrivateKey
 from noise.exceptions import NoiseInvalidMessage
 
 from esphome_device_builder.api.ws import init_ws_app
-from esphome_device_builder.controllers.remote_build import RemoteBuildController
+from esphome_device_builder.controllers.remote_build import OffloaderController
 from esphome_device_builder.controllers.remote_build import _models as rb_models
-from esphome_device_builder.controllers.remote_build import controller as rb
+from esphome_device_builder.controllers.remote_build import offloader as rb
 from esphome_device_builder.controllers.remote_build import (
     peer_link_client as remote_build_peer_link_client,
 )
@@ -100,6 +100,7 @@ from .conftest import (
     capture_events,
     make_remote_build_controller,
 )
+from .conftest import RemoteBuildTestHandles as RemoteBuildController
 
 
 def _make_controller(*, config_dir: Path) -> RemoteBuildController:
@@ -878,7 +879,7 @@ async def test_controller_request_pair_unexpected_status_raises_internal_error(
         )
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller.peer_link_request_pair",
+        "esphome_device_builder.controllers.remote_build.offloader.peer_link_request_pair",
         _fake_request_pair,
     )
 
@@ -1314,7 +1315,7 @@ async def test_start_seeds_pairings_dict_from_disk(
     the spawn side effect.
     """
     monkeypatch.setattr(
-        RemoteBuildController,
+        OffloaderController,
         "_spawn_peer_link_client",
         lambda self, pairing: None,
     )
@@ -1741,7 +1742,7 @@ async def test_request_pair_clears_offloader_alert_for_same_receiver(
         )
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller.peer_link_request_pair",
+        "esphome_device_builder.controllers.remote_build.offloader.peer_link_request_pair",
         _fake_request_pair,
     )
     fake_identity = MagicMock()
@@ -1749,7 +1750,7 @@ async def test_request_pair_clears_offloader_alert_for_same_receiver(
     fake_dashboard = MagicMock()
     fake_dashboard.dashboard_id = "dashboard-stub"
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller._load_offloader_identities",
+        "esphome_device_builder.controllers.remote_build.offloader._load_offloader_identities",
         lambda _config_dir: (fake_identity, fake_dashboard),
     )
     # Park the spawned listener on an unfulfilled wait so the
@@ -1763,7 +1764,7 @@ async def test_request_pair_clears_offloader_alert_for_same_receiver(
         raise AssertionError("park event should never be set in this test")
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller.peer_link_await_pair_status",
+        "esphome_device_builder.controllers.remote_build.offloader.peer_link_await_pair_status",
         _fake_await_pair_status,
     )
 
@@ -1850,7 +1851,7 @@ async def test_request_pair_repair_then_unpair_clean_state(
         return fake_results.pop(0)
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller.peer_link_request_pair",
+        "esphome_device_builder.controllers.remote_build.offloader.peer_link_request_pair",
         _fake_request_pair,
     )
 
@@ -1877,7 +1878,7 @@ async def test_request_pair_repair_then_unpair_clean_state(
         raise AssertionError("park event should never be set in this test")
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller.peer_link_await_pair_status",
+        "esphome_device_builder.controllers.remote_build.offloader.peer_link_await_pair_status",
         _fake_await_pair_status,
     )
     fake_identity = MagicMock()
@@ -1885,7 +1886,7 @@ async def test_request_pair_repair_then_unpair_clean_state(
     fake_dashboard = MagicMock()
     fake_dashboard.dashboard_id = "dashboard-stub"
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller._load_offloader_identities",
+        "esphome_device_builder.controllers.remote_build.offloader._load_offloader_identities",
         lambda _config_dir: (fake_identity, fake_dashboard),
     )
 
@@ -2021,7 +2022,7 @@ async def test_request_pair_repair_against_pending_cancels_old_listener(
         )
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller.peer_link_request_pair",
+        "esphome_device_builder.controllers.remote_build.offloader.peer_link_request_pair",
         _fake_request_pair,
     )
 
@@ -2295,7 +2296,7 @@ async def test_pair_status_listener_loop_backs_off_on_transport_error(
     on the first call and returns APPROVED on the second.
     """
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller._PAIR_STATUS_RECONNECT_BACKOFF_SECONDS",
+        "esphome_device_builder.controllers.remote_build.offloader._PAIR_STATUS_RECONNECT_BACKOFF_SECONDS",
         0.0,
     )
     offloader = _make_offloader_controller(config_dir=offloader_controller_dir)
@@ -2322,7 +2323,7 @@ async def test_pair_status_listener_loop_backs_off_on_transport_error(
         )
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller.peer_link_await_pair_status",
+        "esphome_device_builder.controllers.remote_build.offloader.peer_link_await_pair_status",
         _fake_poll,
     )
     # Stub identity load so it doesn't try to read real key files.
@@ -2331,7 +2332,7 @@ async def test_pair_status_listener_loop_backs_off_on_transport_error(
     fake_dashboard = MagicMock()
     fake_dashboard.dashboard_id = "alpha"
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller._load_offloader_identities",
+        "esphome_device_builder.controllers.remote_build.offloader._load_offloader_identities",
         lambda _config_dir: (fake_identity, fake_dashboard),
     )
 
@@ -2355,7 +2356,7 @@ async def test_pair_status_listener_loop_backs_off_on_unexpected_status(
     tight-loop against it.
     """
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller._PAIR_STATUS_RECONNECT_BACKOFF_SECONDS",
+        "esphome_device_builder.controllers.remote_build.offloader._PAIR_STATUS_RECONNECT_BACKOFF_SECONDS",
         0.0,
     )
     offloader = _make_offloader_controller(config_dir=offloader_controller_dir)
@@ -2385,7 +2386,7 @@ async def test_pair_status_listener_loop_backs_off_on_unexpected_status(
         )
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller.peer_link_await_pair_status",
+        "esphome_device_builder.controllers.remote_build.offloader.peer_link_await_pair_status",
         _fake_poll,
     )
     fake_identity = MagicMock()
@@ -2393,7 +2394,7 @@ async def test_pair_status_listener_loop_backs_off_on_unexpected_status(
     fake_dashboard = MagicMock()
     fake_dashboard.dashboard_id = "alpha"
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.remote_build.controller._load_offloader_identities",
+        "esphome_device_builder.controllers.remote_build.offloader._load_offloader_identities",
         lambda _config_dir: (fake_identity, fake_dashboard),
     )
